Fix LayerHost losing its area by moving it outside of EmbeddedPeoplePane
