+++ packages/react-composites/src/composites/CallWithChatComposite/CallWithChatComposite.tsx
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import React, { useCallback, useState, useMemo, useEffect } from 'react';
-import { PartialTheme, Stack, Theme } from '@fluentui/react';
+import { LayerHost, mergeStyles, PartialTheme, Stack, Theme } from '@fluentui/react';
 import { CallComposite, CallCompositePage, CallControlOptions } from '../CallComposite';
 import { CallAdapterProvider } from '../CallComposite/adapter/CallAdapterProvider';
 import { EmbeddedChatPane } from './EmbeddedChatPane';
@@ -13,7 +13,8 @@ import {
   callCompositeContainerStyles,
   compositeOuterContainerStyles,
   controlBarContainerStyles,
-  drawerContainerStyles
+  drawerContainerStyles,
+  modalLayerHostStyle
 } from './styles/CallWithChatCompositeStyles';
 import { CallWithChatAdapter } from './adapter/CallWithChatAdapter';
 import { CallWithChatBackedCallAdapter } from './adapter/CallWithChatBackedCallAdapter';
@@ -27,6 +28,7 @@ import { ChatAdapterProvider } from '../ChatComposite/adapter/ChatAdapterProvide
 import { CallWithChatAdapterState } from './state/CallWithChatAdapterState';
 import { PreparedMoreDrawer } from './PreparedMoreDrawer';
 import { ParticipantMenuItemsCallback } from '@internal/react-components';
+import { useId } from '@fluentui/react-hooks';
 
 /**
  * Props required for the {@link CallWithChatComposite}
@@ -170,6 +172,8 @@ const CallWithChatScreen = (props: CallWithChatScreenProps): JSX.Element => {
     };
   }, [callWithChatAdapter]);
 
+  const modalLayerHostId = useId('modalLayerhost');
+
   const isInLobbyOrConnecting = currentPage === 'lobby';
   const hasJoinedCall = !!(currentPage && hasJoinedCallFn(currentPage, currentCallState ?? 'None'));
   const showControlBar = isInLobbyOrConnecting || hasJoinedCall;
@@ -214,6 +218,7 @@ const CallWithChatScreen = (props: CallWithChatScreenProps): JSX.Element => {
               onChatButtonClick={selectChat}
               onPeopleButtonClick={selectPeople}
               mobileView={isMobile}
+              modalLayerHostId={modalLayerHostId}
             />
           </CallAdapterProvider>
         )}
@@ -245,6 +250,12 @@ const CallWithChatScreen = (props: CallWithChatScreenProps): JSX.Element => {
           </CallAdapterProvider>
         </ChatAdapterProvider>
       )}
+      {
+        // This layer host is for Modal wrapping the PiPiP in the mobile EmbeddedPeoplePane. This LayerHost can't be inside the EmbeddedPeoplePane
+        // because when the EmbeddedPeoplePane is hidden, ie. style property display is 'none', it takes up no space. This causes problems when dragging
+        // the Modal because the draggable bounds is no space and will always returns to its initial position after dragging.
+        isMobile && <LayerHost id={modalLayerHostId} className={mergeStyles(modalLayerHostStyle)} />
+      }
     </Stack>
   );
 };
